# openapi.yaml
openapi: 3.1.0
info:
  title: Astra Wealth API
  version: 0.1.0
  description: Mock-safe, provider-neutral APIs for Indian wealth intelligence.
servers:
  - url: http://localhost:8000
paths:
...
  /v1/screener/scan:
    post:
      operationId: runScreener
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScreenerRequest' }
      responses:
        '200':
          description: Screener results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScreenerResponse' }
...
components:
  schemas:
...
    ScreenerRequest:
      type: object
      properties:
        universe: { type: string, default: nifty_500 }
        filters: { type: array, items: { type: string } }
        limit: { type: integer, minimum: 1, maximum: 100, default: 25 }
    ScreenerResult:
      type: object
      required: [ticker, company, sector, composite, quality, momentum, valuation]
      properties:
        ticker: { type: string }
        company: { type: string }
        sector: { type: string }
        composite: { type: integer, minimum: 0, maximum: 100 }
        quality: { type: integer, minimum: 0, maximum: 100 }
        momentum: { type: integer, minimum: 0, maximum: 100 }
        valuation: { type: string }
    ScreenerResponse:
      type: object
      required: [query, source_mode, results]
      properties:
        query: { $ref: '#/components/schemas/ScreenerRequest' }
        source_mode: { type: string }
        results: { type: array, items: { $ref: '#/components/schemas/ScreenerResult' } }
...